project/requirements.txt
torch>=1.1
numpy
tqdm
tensorboard>=1.14
torchaudio==0.8.1
torchvision==0.9.1
pandas>=1.2.2
jupyterlab==3.0.9
matplotlib==3.3.4
librosa==0.8.0
mlflow
google-cloud-storage>=1.36.2
dill>=0.3.3
dvc>=2.1.0
invoke>=1.6.0
google-api-python-client>=2.14.0
flake8>=3.9.2